store/requirements.txt
amqp==5.1.1
asgiref==3.7.1
astroid==2.15.5
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-debug-toolbar==4.1.0
django-shopping-cart==0.1
djangorestframework==3.14.0
flake8==6.0.0
flake8-import-order==0.18.2
idna==3.4
kombu==5.3.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
